Simplify sprite concept
We simplify what a sprite is. Normally it points to an atlas for performance reasons, but the aim of this library is really to keep things simple and we don't deal with textures anyway, so there are no benefits to stuff like an Atlas.

Also make make the choice of having Actors deal with floating point positions so that physics calculations work fine. Sprites deal in integer positions however.